edit param counter and rename class AbstractGenerationClassSignature
